sd_server/requirements.txt
torch
diffusers
transformers
accelerate
safetensors
fastapi
uvicorn[standard]
pydantic
python-dotenv
